+++ src/main/java/szewek/mcflux/R.java
@@ -5,11 +5,11 @@ public final class R {
 		MCFLUX_VERSION = "@VERSION@",
 		MCFLUX_NAME = "mcflux",
 		MCFLUX_API_NAME = "mcfluxAPI",
-		MCFLUX_FLAVORAPI_NAME = "mcfluxAPI|flavor",
-		PROXY_SERVER = "szewek.mcflux.proxy.ProxyCommon",
-		PROXY_CLIENT = "szewek.mcflux.proxy.ProxyClient",
-		FORMAT_ENERGY_STAT = "%d / %d",
-		FORMAT_ENERGY_STAT_UNIT = FORMAT_ENERGY_STAT + " MF",
+		MCFLUX_FLAVORAPI_NAME = MCFLUX_API_NAME + "|flavor",
+		MCFLUX_PKG = "szewek.mcflux",
+		PROXY_SERVER = MCFLUX_PKG + ".proxy.ProxyCommon",
+		PROXY_CLIENT = MCFLUX_PKG + ".proxy.ProxyClient",
+		WAILA_REGISTER = MCFLUX_PKG + ".compat.waila.MCFluxWailaProvider.callbackRegister",
 		TAG_MF = "@MF";
 
 	private R() {}