+++ src/main/java/szewek/mcflux/proxy/ProxyCommon.java
@@ -3,5 +3,4 @@ package szewek.mcflux.proxy;
 public class ProxyCommon {
 	public void preInit() {}
 	public void init() {}
-	public void postInit() {}
 }